requirements.txt
# ServicioEjecucion - Requirements consolidados
# Incluye dependencias de todos los scrapers + ProductUploader

# MongoDB y configuración
pymongo==4.6.0
python-dotenv==1.0.0

# Web scraping común (alkosto, exito)
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.2

# Selenium y Playwright (falabella - opcional por ahora)
selenium==4.35.0
webdriver-manager==4.0.2
playwright==1.54.0

# Dependencias de Selenium/Playwright
attrs==25.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
greenlet==3.2.4
h11==0.16.0
idna==3.10
outcome==1.3.0.post0
packaging==25.0
pycparser==2.22
pyee==13.0.0
PySocks==1.7.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0